Summary
threw the files into proper github folder and finally added them to github.

makepassword.js
 - refactored to give more errors in case of wrong inputs

passwordjs.js
 - wrote to check if the passed in email and passwords exists within the file.

utility.js
 - pulled 'checkExists' out of makepassword and added it here to use across more of the files

makepassword.test.js
 - wrote the tests.
- played around with webstorm's suggested improvements and streamlining
- it is... over eager with the tab help, but sometimes it gets it right.
- also nice to have contextual tags on what variables are expected, what type the variables are expected to be.

+++ src/makepassword.js
@@ -1,40 +1,69 @@
 'use strict'
 const fs = require('fs');
-const {readFile, writeFile, hash} = require('./utility')
+const {readFile, writeFile, hash, checkExists} = require('./utility')
 
 function makepassword(passwordFileName, passwordEncFileName) {
     // read file in. File will be returned as an array of lines.
     let inputValues = [];
     let outputValues = [];
-    
-    inputValues = readFile(passwordFileName)
-    console.log(inputValues)
+    let inputFileExists = ''
+    let outputFileExists = ''
 
-    // split the data further on the : to get passwords and convert them.
-    inputValues.forEach(encodePassword)
-    
-    // Convert passwords to encrypted, call Hash in utility.
-    function encodePassword(item) {
-        let encodedPassword = '';
-        let splitArray = [];
-        
-        // splitting original item to get the password
-        splitArray = item.split(":");
-        
-        // encoding the password
-        encodedPassword = hash(splitArray[1]);
+    inputFileExists = checkExists(passwordFileName);
 
-        // Writing the values to the output array
-        outputValues.push(splitArray[0] + ":" + encodedPassword)
+    console.log("File status: " + inputFileExists);
+    if (inputFileExists === 'Found'){
+        try{
+            inputValues = readFile(passwordFileName)
+        } catch {
+            console.log("Error reading file!")
+        }
+
+        // split the data further on the : to get passwords and convert them.
+        inputValues.forEach(encodePassword)
+
+        // Convert passwords to encrypted, call Hash in utility.
+        // Declared here for scoping.
+        function encodePassword(item) {
+            let encodedPassword = '';
+            let splitArray = [];
+
+            // splitting original item to get the password
+            splitArray = item.split(":");
+            try {
+                // encoding the password
+                encodedPassword = hash(splitArray[1]);
+
+                // Writing the values to the output array
+                outputValues.push(splitArray[0] + ":" + encodedPassword)
+
+            } catch {
+                console.log("Error hashing the passwords!")
+            }
+        }
+
+        // check if encrypted file exists
+        outputFileExists = checkExists(passwordFileName);
+
+        if (outputFileExists === 'Found') {
+            try {
+
+            } catch {
+
+            }
+        } else {
+            // file not found, we can write it.
+           try {
+               // write encrypted out to file.
+               writeFile(outputValues, passwordEncFileName);
+           } catch {
+               console.log("Error writing to " + passwordFileName);
+           }
+        }
+
+    } else {
+        console.log( "No file to read. Please try running the program gain.")
     }
-    console.log(outputValues)
-
-    // write encrypted out to file.
-    writeFile(outputValues, passwordEncFileName);
-
-    // write encrypted data to mongoDB
-    
-
 }
 
 if (require.main === module) {
